+++ src/components/Form.tsx
@@ -4,17 +4,30 @@ import { StyledForm } from '@styles/components/Form.styled';
 
 type FormProps = {
   children: React.ReactNode;
+  handleLabelClick: (label: string) => void;
+  isSignup: boolean;
   onSubmit: () => void;
 };
 
 type FormComposition = { Input?: React.FC<InputProps> };
 
 const Form: React.FC<FormProps> & FormComposition =
-  ({ children, onSubmit }: FormProps): JSX.Element =>
+  ({ children, onSubmit, handleLabelClick, isSignup }: FormProps): JSX.Element =>
     <StyledForm>
-      <label>LOGIN</label>
-      { children }
-      <input onClick={ onSubmit } type="submit" value="SUBMIT" />
+      <label
+        className={`label-login ${!isSignup ? 'active' : ''}`}
+        onClick={(): void => handleLabelClick('login')}
+      >
+        LOGIN
+      </label>
+      <label
+        className={`label-signup ${isSignup ? 'active' : ''}`}
+        onClick={(): void => handleLabelClick('signup')}
+      >
+        SIGN UP
+      </label>
+      {children}
+      <input onClick={onSubmit} type='submit' value={isSignup ? 'SIGN UP' : 'LOGIN'} />
     </StyledForm>;
 
 export default Form;